introducing concrete integer types

+++ parser/utils/parse_num.h
@@ -65,6 +65,6 @@ class Value;
 class ASTAllocator;
 class TypeBuilder;
 
-Value* allocate_number_value(ASTAllocator& alloc, TypeBuilder& typeBuilder, unsigned long long value, SourceLocation location);
+Value* allocate_number_value(ASTAllocator& alloc, TypeBuilder& typeBuilder, uint64_t value, SourceLocation location);
 
 parse_num_result<Value*> convert_number_to_value(ASTAllocator& alloc, TypeBuilder& typeBuilder, const char* value, std::size_t value_size, bool is64Bit, SourceLocation location);
